+++ source/Public/Get-LocalizedData.ps1
@@ -1,3 +1,4 @@
+
 <#
     .SYNOPSIS
         Gets language-specific data into scripts and functions based on the UI culture
@@ -175,7 +176,11 @@ function Get-LocalizedData
             Because Proxy Command changes the Invocation origin, we need to be explicit
             when handing the pipeline back to original command.
         #>
-        if (!$PSBoundParameters.ContainsKey('FileName'))
+        if ($PSBoundParameters.ContainsKey('FileName'))
+        {
+            Write-Debug -Message ('Looking for provided file with base name: ''{0}''.' -f $FileName)
+        }
+        else
         {
             if ($myInvocation.ScriptName)
             {
@@ -189,6 +194,7 @@ function Get-LocalizedData
             $FileName = $file.BaseName
 
             $PSBoundParameters.Add('FileName', $file.Name)
+            Write-Debug -Message ('Looking for resolved file with base name: ''{0}''.' -f $FileName)
         }
 
         if ($PSBoundParameters.ContainsKey('BaseDirectory'))
@@ -198,75 +204,81 @@ function Get-LocalizedData
         else
         {
             $callingScriptRoot = $MyInvocation.PSScriptRoot
-
             $PSBoundParameters.Add('BaseDirectory', $callingScriptRoot)
         }
 
-        if ($PSBoundParameters.ContainsKey('DefaultUICulture') -and !$PSBoundParameters.ContainsKey('UICulture'))
+        # If we're not looking for a specific UICulture, but looking for current culture, one of its parent, or the default.
+        if (-not $PSBoundParameters.ContainsKey('UICulture') -and $PSBoundParameters.ContainsKey('DefaultUICulture'))
         {
             <#
                 We don't want the resolution to eventually return the ModuleManifest
                 so we run the same GetFilePath() logic than here:
                 https://github.com/PowerShell/PowerShell/blob/master/src/Microsoft.PowerShell.Commands.Utility/commands/utility/Import-LocalizedData.cs#L302-L333
-                and if we see it will return the wrong thing, set the UICulture to DefaultUI culture, and return the logic to Import-LocalizedData
-            #>
-            $currentCulture = Get-UICulture
+                and if we see it will return the wrong thing, set the UICulture to DefaultUI culture, and return the logic to Import-LocalizedData.
 
+                If the LCID is 127 (invariant) then use default UI culture anyway.
+                If we can't create the CultureInfo object, it's probably because the Globalization-invariant mode is enabled for the DotNet runtime (breaking change in .Net)
+                See more information in issue https://github.com/dsccommunity/DscResource.Common/issues/11.
+                https://docs.microsoft.com/en-us/dotnet/core/compatibility/globalization/6.0/culture-creation-invariant-mode
+            #>
+
+            $currentCulture = Get-UICulture
             $evaluateDefaultCulture = $true
 
-            <#
-                If the LCID is 127 then use default UI culture instead.
-
-                See more information in issue https://github.com/dsccommunity/DscResource.Common/issues/11.
-            #>
             if ($currentCulture.LCID -eq 127)
             {
-                $currentCulture = New-Object -TypeName 'System.Globalization.CultureInfo' -ArgumentList @($DefaultUICulture)
-                $PSBoundParameters['UICulture'] = $DefaultUICulture
+                try
+                {
+                    # Current culture is invariant, let's directly evaluate the DefaultUICulture
+                    $currentCulture = New-Object -TypeName 'System.Globalization.CultureInfo' -ArgumentList @($DefaultUICulture)
+                    # No need to evaluate the DefaultUICulture later, as we'll start with this (in the while loop below)
+                    $evaluateDefaultCulture = $false
+                }
+                catch
+                {
+                    Write-Debug -Message 'The Globalization-Invariant mode is enabled, only the Invariant Culture is allowed.'
+                    # The code will now skip to the InvokeCommand part and execute the Get-LocalizedDataForInvariantCulture
+                }
 
-                $evaluateDefaultCulture = $false
+                $PSBoundParameters['UICulture'] = $DefaultUICulture
             }
 
-            $languageFile = $null
-
-            $localizedFileNames = @(
-                $FileName + '.psd1'
-                $FileName + '.strings.psd1'
+            [string] $languageFile = ''
+            [string[]] $localizedFileNamesToTry = @(
+                ('{0}.psd1' -f $FileName)
+                ('{0}.strings.psd1' -f $FileName)
             )
 
-            while ($null -ne $currentCulture -and $currentCulture.Name -and -not $languageFile)
+            while (-not [string]::IsNullOrEmpty($currentCulture.Name) -and [String]::IsNullOrEmpty($languageFile))
             {
-                foreach ($fullFileName in $localizedFileNames)
+                Write-Debug -Message ('Looking for Localized data file using the current culture ''{0}''.' -f $currentCulture.Name)
+                foreach ($localizedFileName in $localizedFileNamesToTry)
                 {
-                    $filePath = [System.IO.Path]::Combine($callingScriptRoot, $CurrentCulture.Name, $fullFileName)
-
+                    $filePath = [System.IO.Path]::Combine($callingScriptRoot, $CurrentCulture.Name, $localizedFileName)
                     if (Test-Path -Path $filePath)
                     {
-                        Write-Debug -Message "Found $filePath"
-
+                        Write-Debug -Message "Found '$filePath'."
                         $languageFile = $filePath
-
                         # Set the filename to the file we found.
-                        $PSBoundParameters['FileName'] = $fullFileName
-
-                        # Exit loop if we find the first filename.
+                        $PSBoundParameters['FileName'] = $localizedFileName
+                        # Exit loop if as we found the first filename.
                         break
                     }
                     else
                     {
-                        Write-Debug -Message "File $filePath not found"
+                        Write-Debug -Message "File '$filePath' not found."
                     }
                 }
 
-                if (-not $languageFile)
+                if ([String]::IsNullOrEmpty($languageFile))
                 {
                     <#
-                        Evaluate the parent culture if there is one.
+                        Evaluate the parent culture if there is a valid one (not Invariant).
 
                         If the parent culture is LCID 127 then move to the default culture.
                         See more information in issue https://github.com/dsccommunity/DscResource.Common/issues/11.
                     #>
-                    if ($currentCulture.Parent -and $currentCulture.Parent.LCID -ne 127)
+                    if ($currentCulture.Parent -and [string]$currentCulture.Parent.Name)
                     {
                         $currentCulture = $currentCulture.Parent
                     }
@@ -281,7 +293,18 @@ function Get-LocalizedData
                                 system UI culture. Evaluating the default UI culture (which
                                 defaults to 'en-US' if not specifically set).
                             #>
-                            $currentCulture = New-Object -TypeName 'System.Globalization.CultureInfo' -ArgumentList @($DefaultUICulture)
+                            try
+                            {
+                                $currentCulture = New-Object -TypeName 'System.Globalization.CultureInfo' -ArgumentList @($DefaultUICulture)
+                            }
+                            catch
+                            {
+                                Write-Debug -Message ('Unable to create the Default UI Culture [CultureInfo] object, most likely due to invariant mode being enabled.')
+                                $currentCulture = Get-UICulture
+                                # We already tried everything we could, exit the while loop and hand over to Import-LocalizedData or Get-LocalizedDataForInvariantCultureMode
+                                break
+                            }
+
                             $PSBoundParameters['UICulture'] = $DefaultUICulture
                         }
                         else
@@ -312,8 +335,25 @@ function Get-LocalizedData
                 $PSBoundParameters['OutBuffer'] = 1
             }
 
-            $wrappedCmd = $ExecutionContext.InvokeCommand.GetCommand('Microsoft.PowerShell.Utility\Import-LocalizedData', [System.Management.Automation.CommandTypes]::Cmdlet)
-            $scriptCmd = { & $wrappedCmd @PSBoundParameters }
+            if ($currentCulture.LCID -eq 127)
+            {
+                # Culture is invariant, working around issue with Import-LocalizedData when pwsh configured as invariant
+                $wrappedCmd = $ExecutionContext.InvokeCommand.GetCommand('Get-LocalizedDataForInvariantCulture', [System.Management.Automation.CommandTypes]::Function)
+                $PSBoundParameters.Keys.ForEach({
+                    if ($_ -notin $wrappedCmd.Parameters.Keys)
+                    {
+                        $PSBoundParameters.Remove($_)
+                    }
+                })
+
+                $scriptCmd = { & $wrappedCmd @PSBoundParameters }
+            }
+            else
+            {
+                <# Action when all if and elseif conditions are false #>
+                $wrappedCmd = $ExecutionContext.InvokeCommand.GetCommand('Microsoft.PowerShell.Utility\Import-LocalizedData', [System.Management.Automation.CommandTypes]::Cmdlet)
+                $scriptCmd = { & $wrappedCmd @PSBoundParameters }
+            }
 
             $steppablePipeline = $scriptCmd.GetSteppablePipeline($myInvocation.CommandOrigin)
             $steppablePipeline.Begin($PSCmdlet)